+++ CHANGELOG.md
@@ -65,6 +65,7 @@ v2.0.0
 - Removed mixed modes for `String#hankaku` and `String#zenkaku`.
 - Modified `String#replaceTags` and `String#stripTags` to simplify argument signature and small performance boost.
 - Added `Array#isEqual`.
+- Renamed `RegExp#addFlag` and `RegExp#removeFlag` to `RegExp#addFlags` and `RegExp#removeFlags`, and ensured they work on multiple flags.
 
 
 v1.4.2